backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

langgraph==0.4.0
langchain==0.3.0
langchain-core==0.3.0
langchain-google-genai==4.2.1

python-dotenv==1.0.1
pydantic==2.8.2
typing_extensions==4.12.2